# Replace numpy transpose with torch permute to speed-up