--- a/mmdet/datasets/pipelines/formatting.py
+++ b/mmdet/datasets/pipelines/formatting.py
@@ -237,9 +237,7 @@ class DefaultFormatBundle:
             else:
                 img = to_tensor(img).permute(2, 0, 1).contiguous()
             results['img'] = DC(
-                img,
-                padding_value=self.pad_val['img'],
-                stack=True)
+                img, padding_value=self.pad_val['img'], stack=True)
         for key in ['proposals', 'gt_bboxes', 'gt_bboxes_ignore', 'gt_labels']:
             if key not in results:
                 continue