--- a/requirements/optional.txt
+++ b/requirements/optional.txt
@@ -1,3 +1,3 @@
 cityscapesscripts
 imagecorruptions
-sklearn
+scikit-learn