## Motivation

`numpy.transpose()` is mush more slow than `torch.permute()` according to my benchmarks on a jupyter notebook:

1. If the input image size (1366, 800, 3) is the size of COCO dataset image

```
import numpy as np
import torch
img = np.random.randn(1366, 800, 3)
```

```
%%timeit
img_np = np.ascontiguousarray(img.transpose(2, 0, 1))
input = torch.from_numpy(img_np)
```
Output: 7.69 ms ± 314 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

```
%%timeit
input = torch.from_numpy(img).permute(2, 0, 1).contiguous()
```
Output: 1.65 ms ± 123 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

2. If the input image size is large, especially when inferencing a large image ((3648, 5472, 3) in my case)

```
img = np.random.randn(3648, 5472, 3)
```

```
%%timeit
img_np = np.ascontiguousarray(img.transpose(2, 0, 1))
input = torch.from_numpy(img_np)
```
Output: 327 ms ± 1.13 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```
%%timeit
input = torch.from_numpy(img).permute(2, 0, 1).contiguous()
```
Output: 93.8 ms ± 4.77 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

## Modification

Replace the transpose operation`numpy.transpose(2, 0, 1)` with `torch.permute(2, 0, 1)` to in [ImageToTensor](https://github.com/open-mmlab/mmdetection/blob/master/mmdet/datasets/pipelines/formatting.py#L96) and [DefaultFormatBundle](https://github.com/open-mmlab/mmdetection/blob/master/mmdet/datasets/pipelines/formatting.py#L229-L230) to speed-up the process.