@ZwwWayne 2022-12-27
Hi @Min-Sheng , 
Thanks for your kind PR. Would you like to also update the docstring or code to indicate this issue？