@Min-Sheng 2022-12-27
> Hi @Min-Sheng , Thanks for your kind PR. Would you like to also update the docstring or code to indicate this issue？

I have updated both the docstring and code.
By the way, I found that if the input numpy array is non-contiguous,
```
img = np.random.randn(1366, 800, 3)
img = img[..., ::-1]
```
use
```
input = torch.from_numpy(np.ascontiguousarray(img.transpose(2, 0, 1)))
```
Output: 7.58 ms ± 118 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

is faster than
```
input = torch.from_numpy(np.ascontiguousarray(img)).permute(2, 0, 1).contiguous()
```
Output: 14.5 ms ± 669 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

So, I use the numpy c_contiguous flag for array continuousness to switch the order of transpose and to_tensor operations.